Update gui-backup.txt
This is a Python script that creates a graphical user interface (GUI) using the PyQt5 library for controlling a scanner. The GUI includes buttons for starting and stopping the scanner, adjusting LED brightness, and returning the scanner to its home position. The script also includes code for initializing the neopixels and cleaning up the GPIO pins on exit. The scanner functionality is defined in the scanner module. The GUI elements are organized using a combination of QVBoxLayout and QHBoxLayout, and customized QSpinBox and QSlider classes are used to modify the appearance of input fields. The script is designed to run on a Raspberry Pi with a 7-inch touchscreen display.